split pokedex file
